Move getPostTypeStaticProps to functions
