+++ pages/404.js
@@ -1,7 +1,7 @@
 import Container from '@/components/atoms/Container'
 import Layout from '@/components/common/Layout'
 import getPagePropTypes from '@/functions/getPagePropTypes'
-import getPostTypeStaticProps from '@/lib/wordpress/_global/getPostTypeStaticProps'
+import getPostTypeStaticProps from '@/functions/wordpress/postTypes/getPostTypeStaticProps'
 import Page from './[...slug]'
 
 // Define route post type.
