+++ pages/login.js
@@ -3,7 +3,7 @@ import Text from '@/components/atoms/Inputs/Text'
 import RichText from '@/components/atoms/RichText'
 import Layout from '@/components/common/Layout'
 import Form from '@/components/molecules/Form'
-import getPostTypeStaticProps from '@/lib/wordpress/_global/getPostTypeStaticProps'
+import getPostTypeStaticProps from '@/functions/wordpress/postTypes/getPostTypeStaticProps'
 import {signIn, useSession} from 'next-auth/client'
 import Link from 'next/link'
 import {useRouter} from 'next/router'
